Master/Worker mode for summarize_mpi.py
Allow the tasks to be handed out as nodes finish the previous one.
Should allow for better packing.

Also, include a flag to use a library version of pmlogextract.  Suggest
usage of this when running in MPI mode since forking from an MPI task
has shown to cause segfaults in some MPI implementations.

Change the logging for mpi jobs.  Each task has its own log file. printf
from the dlopened pypmlogextract will not be captured.  That fix is a WIP.