Revert setuptools changes.
The changes to the setuptools config that was done to enable
compatibility with python nose tests broke the excutuable scripts that
are installed. This reverts these changes and (temporarily) removes the
unit testing from running automatically in the CI build. An exra step in
the CI build has been added to check that the scripts can execute.