Do not automatically update to latest.
The docker that runs in shippable does not support the cap_set_file
API that is needed by the httpd RPM. This means that it is not
possible to update the httpd RPM in shippable.

The plan going forward is to manage the software versions in
the 'FROM' docker image.