SlurmProc preprocessor performance tuning (#147)
* Precompile regex for slurmcgroupparser instead of compiling each call

* Do re.search over whole slurm cgroup str instead of splitting by ";"

The split is unnecessary since the first match is returned anyways.
Changing the "^" anchor to "(^|;)" and using search (which finds
anywhere in the string) instead of match (which automatically anchors
at the beginning) keeps the same behavior but improves performance.

* Simplify string operations to find command in SlurmProc

Instead of splitting on space and joining the tail of that list
on space, just split once on space and take the second element of that
list.

* Use faster "<expected> in <str>" instead of <str>.find != -1

* Remove the (^|;) anchor from slurm cgroup parsing regex

This increases the speed of the regex search by nearly 2x, due to
optimizations the c implementation of python regexes perform when the
pattern has a constant string prefix. This will match the same place in
all normal cases.

* Fix some "== None" comparisons to follow pep8

* Slice string to get command instead of splitting

Measurements show this is slightly faster due to not having to copy
the string and create a list object.