Fix cython build and import process (hopefully) for good this time
Use the recommended cythonize to compile extensions instead
of setuptools build_ext. Fix module names of puffypcp and pypmlogextract
(the module is puffypcp, in package supremm.puffypcp) and add
__init__.py files so they're recognized as packages and .so files get
put in the right places. Change imports of puffypcp to match this. Also
change the name of the pcp.pxd "header" file so it doesn't get confused
with the python pcp wrapper (unsure if this is actually necessary but
it's probably a good idea either way).