Refactor summarizejob to separate out database concerns
Return results of summarization instead of writing to databases
within the method. This will allow more flexibility with multiprocessing
and threading, and has the side effect of simplifying the unit tests
(don't need to mock db objects and intercept calls). Top-level exception
handling and timing are also done outside the summarizejob method.