indexarchives speedup (#165)
* Add faster archive indexing method to indexarchives

Avoids creating pcp contexts for job-level archives, instead parsing
the job start timestamp from the filename. Must be enabled by a config
file option, and archive files must be in the correctly structured
directory format.

* Extract db insert into a separate step in the indexing pipeline

Create a contextmanager class for the archive cache.

* Modify db insert to load data from files instead of row by row

This increases speed of insert dramatically. Also adds optional
multiprocessing version of the file parsing loop, this seems to have
higher throughput.

* Add insert_from_files to ArchiveCache abstract class

Raise not implemented in DbArchiveCache for now

* Add indexarchvies opts for number of processes

* Be more explicit about the tri-state logic in archive finder

* Remove ArchiveIndexUpdater, the load file version will be the only option

* Add pytz and tzlocal to rpm required packages

* Remove some debug code from indexarchives

* Add insert_from_files to ArchiveCache abstract class

Raise not implemented in DbArchiveCache for now

* Delete csv files by default and add option to keep for debugging

* Add new indexarchive config options to example config file with comments