Continue processing archives that have PM_ERR_VALUE errors
The original code would give up processing an archive if it saw a metric that
returned PM_ERR_VALUE at any time point. This error is seen with the Ipmi pmda
and is not fatal. If this error is seen it ususaly occurs a few times a day.

This change make the framework skip the timestep that has the error but it will
continue to process the archive and give callbacks for subsequent data.