Restrict job start end PCP archives by time.
The current code always includes the start and end archives when processing a
job. These archives are selected based on the resource, hostname and
local_job_id. This is an issue on resources where the local_job_id is not a unique identifier
for the job*. In this scenario the job start and end archives for all jobs with the
same hostname and local_job_id are processed.

This change resricts the time window of job archives to be within 30 seconds of
the job start/end as reported by the resource manager. The 30 second value was chosen
as it is the default recommended periodic collection period.

* For example this can happen if a job is requeued due to a node fail or if the
local_job_id is reset by a system administrator.