Add support for specifying the meaning of the job script datestamp.
The jobs scripts are stored in datestamped directories. The original
code used the datestamp as the job's start time when looking up the job
information in the database. This change allows the datestamp to be
interpreted as either the submit day, start day, end day or for
the datestamp to be ignored completely.