Only open database connection when needed.
If the pcp archive indexing took longer than the database
connection timeout then the indexarchives script would hang when
trying to do the database write. I would have assumed that the
database connection library would throw a timeout exception, but
this is not seen and the software just hangs.

This change just creates the database connection just before it
is needed for the write.