Fix another edge case error handling in the Performance Counter plugin.
Handle case where a subset of the perf counters had erroneous values. An example
of when this can happen is if some of the counters are multiplexed and
do not have a high enough duty cycle to produce accurate data.

This one would have been caught in 1.4.0 but missed in 1.4.1.