Enable autocommit for markasdone.
This enabled autocommit for the sql queries that updates the XDMoD
table that records jobs that have been processed. This is to mitigate a
problem observed when a connection times out: the code tries to rerun the
queries, but gets a lock wait timout as the first attempt is waint to be
committed (and can't becouase the connection has timedout).