Fix encoding of string data from PCP.
I've been working with the latest SUPREMM beta on RHEL 8, and have found a bug that was causing missing/erroneous data on the Job Viewer page under the Executable Information tab.

The root cause of the problem is that when the module is pulling strings from PCP via pmlogextract, it's getting them as byte arrays, and apparently (at least) with python3.6, using str() on byte array "foo" encodes the resulting string as the literal "b'foo'".  This errant string then doesn't match the string "foo" and causes the missing data.

There are minor fixes to:
src/supremm/pcpcinterface/pcpcinterface.pyx
replacing the three occurrences of str(foo) with str(foo, 'utf-8')

and a fix to src/supremm/preprocessors/Proc.py
to remove a now unnecessary str() call due to the above change.